changing data type of price

+++ src/main/kotlin/com/kiranaStore/api/model/Product.kt
@@ -7,5 +7,5 @@ data class Product(
         val id: Int,
         val name: String,
         val description: String,
-        val price: BigDecimal,
+        val price: Double,
         val status: String) {}
